+++ packages/insights-charts/styleguide/js/AreaCharts.js
@@ -5,12 +5,13 @@ import { getRandomData, getRandomCategories } from './helpers';
 import ReflectChart from '../../source/js/ReflectCharts';
 
 const propTypes = {
+  animations: PropTypes.bool.isRequired,
   sparseness: PropTypes.number.isRequired,
 };
 
 class AreaCharts extends React.Component {
   updateCharts() {
-    const { sparseness } = this.props;
+    const { animations, sparseness } = this.props;
     const dataPoints = 10;
     const dataArr = getRandomData(dataPoints, { sparseness });
 
@@ -30,6 +31,9 @@ class AreaCharts extends React.Component {
     };
 
     const options = {
+      animations: {
+        enabled: animations,
+      },
       area: {
         line: {
           enabled: false,
@@ -125,7 +129,8 @@ class AreaCharts extends React.Component {
 AreaCharts.propTypes = propTypes;
 
 const mapStateToProps = state => ({
-  sparseness: state.options.sparseness || 0,
+  animations: state.options.animations,
+  sparseness: state.options.sparseness,
 });
 
 export default connect(mapStateToProps)(AreaCharts);