+++ packages/insights-charts/styleguide/js/ColumnCharts.js
@@ -5,12 +5,13 @@ import { getRandomData, getRandomCategories } from './helpers';
 import ReflectChart from '../../source/js/ReflectCharts';
 
 const propTypes = {
+  animations: PropTypes.bool.isRequired,
   sparseness: PropTypes.number.isRequired,
 };
 
 class ColumnCharts extends React.Component {
   updateCharts() {
-    const { sparseness } = this.props;
+    const { animations, sparseness } = this.props;
     const dataPoints = 10;
 
     const data = {
@@ -38,6 +39,9 @@ class ColumnCharts extends React.Component {
     };
 
     const options = {
+      animations: {
+        enabled: animations,
+      },
       column: {
         data_labels: {
           enabled: true,
@@ -148,7 +152,8 @@ class ColumnCharts extends React.Component {
 ColumnCharts.propTypes = propTypes;
 
 const mapStateToProps = state => ({
-  sparseness: state.options.sparseness || 0,
+  animations: state.options.animations,
+  sparseness: state.options.sparseness,
 });
 
 export default connect(mapStateToProps)(ColumnCharts);