+++ packages/insights-charts/styleguide/js/CombinationCharts.js
@@ -5,12 +5,13 @@ import { getRandomData, getRandomCategories } from './helpers';
 import ReflectChart from '../../source/js/ReflectCharts';
 
 const propTypes = {
+  animations: PropTypes.bool.isRequired,
   sparseness: PropTypes.number.isRequired,
 };
 
 class CombinationCharts extends React.Component {
   updateCharts() {
-    const { sparseness } = this.props;
+    const { animations, sparseness } = this.props;
     const dataPoints = 10;
 
     const data = {
@@ -37,6 +38,9 @@ class CombinationCharts extends React.Component {
     };
 
     const options = {
+      animations: {
+        enabled: animations,
+      },
       column: {
         layout: 'grouped',
         data_labels: {
@@ -119,7 +123,8 @@ class CombinationCharts extends React.Component {
 CombinationCharts.propTypes = propTypes;
 
 const mapStateToProps = state => ({
-  sparseness: state.options.sparseness || 0,
+  animations: state.options.animations,
+  sparseness: state.options.sparseness,
 });
 
 export default connect(mapStateToProps)(CombinationCharts);