+++ packages/insights-charts/styleguide/js/ScatterCharts.js
@@ -5,12 +5,13 @@ import { getRandomData, getRandomCategories } from './helpers';
 import ReflectChart from '../../source/js/ReflectCharts';
 
 const propTypes = {
+  animations: PropTypes.bool.isRequired,
   sparseness: PropTypes.number.isRequired,
 };
 
 class ScatterCharts extends React.Component {
   updateCharts() {
-    const { sparseness } = this.props;
+    const { animations, sparseness } = this.props;
     const dataPoints = 10;
 
     const multiData = {
@@ -39,7 +40,7 @@ class ScatterCharts extends React.Component {
 
     const options = {
       animations: {
-        enabled: false,
+        enabled: animations,
       },
       axis: {
         x: {},
@@ -139,7 +140,8 @@ class ScatterCharts extends React.Component {
 ScatterCharts.propTypes = propTypes;
 
 const mapStateToProps = state => ({
-  sparseness: state.options.sparseness || 0,
+  animations: state.options.animations,
+  sparseness: state.options.sparseness,
 });
 
 export default connect(mapStateToProps)(ScatterCharts);